+++ fm_eventmanager/settings.py.devel
@@ -87,6 +87,7 @@ INSTALLED_APPS = [
     'django_u2f',
     'maintenance_mode',
     'registration',
+    'django_vite',
 ]
 
 MIDDLEWARE = [
@@ -249,3 +250,5 @@ ENVIRONMENT_NAME = "Production Server"
 ENVIRONMENT_COLOR = "#FF0000"
 ENVIRONMENT_TEXT_COLOR = "#00FF00"
 ENVIRONMENT_FLOAT = True
+
+DJANGO_VITE_STATIC_URL_PREFIX = "bundler"
