+++ fm_eventmanager/settings.py.example
@@ -55,6 +55,7 @@ INSTALLED_APPS = [
     'maintenance_mode',
     'registration',
     'debug_toolbar',
+    'django_vite',
 ]
 
 MIDDLEWARE = [
@@ -207,3 +208,6 @@ ENVIRONMENT_NAME = "Production Server"
 ENVIRONMENT_COLOR = "#FF0000"
 ENVIRONMENT_TEXT_COLOR = "#00FF00"
 ENVIRONMENT_FLOAT = True
+
+DJANGO_VITE_DEV_MODE = DEBUG
+DJANGO_VITE_STATIC_URL_PREFIX = "bundler"
