+++ fm_eventmanager/settings.py.travis
@@ -86,6 +86,7 @@ INSTALLED_APPS = [
     'maintenance_mode',
     'registration',
     'debug_toolbar',
+    'django_vite',
 ]
 
 MIDDLEWARE = [
@@ -244,3 +245,5 @@ ENVIRONMENT_TEXT_COLOR = "#00FF00"
 ENVIRONMENT_FLOAT = True
 
 GOTENBERG_HOST = "http://localhost:3000"
+
+DJANGO_VITE_STATIC_URL_PREFIX = "bundler"
