+++ registration/frontend/package.json
@@ -1,21 +1,30 @@
 {
+  "name": "apis-frontend",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc -b && vite build"
+  },
   "dependencies": {
     "@fortawesome/fontawesome-free": "^7.1.0",
     "@kobalte/core": "^0.13.11",
-    "@sentry/solid": "^10.20.0",
+    "@sentry/solid": "^10.25.0",
     "@solid-primitives/keyboard": "^1.3.3",
     "big.js": "^7.0.1",
     "bulma": "^1.0.4",
     "date-fns": "^4.1.0",
-    "esbuild": "^0.25.11",
-    "esbuild-plugin-solid": "^0.6.0",
-    "esbuild-sass-plugin": "^3.3.1",
     "mitt": "^3.0.1",
     "mqtt": "^5.14.1",
-    "solid-js": "^1.9.9"
+    "solid-js": "^1.9.10"
   },
   "devDependencies": {
     "@types/big.js": "^6.2.2",
-    "typescript": "^5.9.3"
+    "prettier": "^3.6.2",
+    "prettier-plugin-organize-imports": "^4.3.0",
+    "sass-embedded": "^1.93.3",
+    "typescript": "^5.9.3",
+    "vite": "^7.2.2",
+    "vite-plugin-solid": "^2.11.10"
   }
 }
