+++ registration/frontend/src/admin/attendee-search/components/AttendeeSearch.tsx
@@ -1,19 +1,19 @@
 import { createShortcut } from "@solid-primitives/keyboard";
 import {
-  Accessor,
-  Component,
+  type Accessor,
+  type Component,
   createEffect,
   createResource,
   createSignal,
   For,
-  Setter,
+  type Setter,
   Show,
   untrack,
   useContext,
 } from "solid-js";
 
-import { BadgeResult, getSearchResults } from "..";
-import { SentryErrorBoundary } from "../../../entrypoints/admin";
+import { type BadgeResult, getSearchResults } from "..";
+import { SentryErrorBoundary } from "../..";
 import { CartManager } from "../../cart";
 import { ConfigContext } from "../../providers/config-provider";
 import { BadgeTableLoader } from "./BadgeTableLoader";
@@ -65,7 +65,7 @@ export const AttendeeSearch: Component<{
 
   const [results, { refetch }] = createResource(
     props.searchQuery,
-    async (query) => await getSearchResults(config.urls, query)
+    async (query) => await getSearchResults(config.urls, query),
   );
 
   let searchInputRef!: HTMLInputElement;
@@ -102,7 +102,7 @@ export const AttendeeSearch: Component<{
 
       const firstResultHasExact = hasAnyExactMatch(
         cleanedSearchQuery,
-        entries[0]
+        entries[0],
       );
       const anyOtherResultHasExact =
         firstResultHasExact &&
@@ -144,7 +144,7 @@ export const AttendeeSearch: Component<{
         }
       } else {
         const next = entries.find(
-          (badge) => !props.cartManager.alreadyInCart(badge.id)
+          (badge) => !props.cartManager.alreadyInCart(badge.id),
         );
         if (next) {
           props.cartManager.addCartId(next.id);
@@ -169,7 +169,7 @@ export const AttendeeSearch: Component<{
     },
     {
       preventDefault: true,
-    }
+    },
   );
 
   const noResults = (
