+++ registration/frontend/src/admin/attendee-search/index.ts
@@ -1,4 +1,4 @@
-import { ApisUrls, CSRF_TOKEN } from "../../entrypoints/admin";
+import { type ApisUrls, CSRF_TOKEN } from "..";
 import { AttendeeSearch } from "./components/AttendeeSearch";
 
 export { AttendeeSearch };
@@ -20,7 +20,7 @@ export interface Attendee {
 
 export async function getSearchResults(
   urls: ApisUrls,
-  query: string
+  query: string,
 ): Promise<BadgeResult[]> {
   // Clear results if we search for an empty string.
   if (query.trim().length === 0) {
