+++ registration/frontend/src/admin/cart/cart-manager.ts
@@ -1,6 +1,6 @@
-import { Accessor, createSignal, Setter } from "solid-js";
+import { type Accessor, createSignal, type Setter } from "solid-js";
 
-import { ApisUrls, CSRF_TOKEN } from "../../entrypoints/admin";
+import { type ApisUrls, CSRF_TOKEN } from "..";
 import MqttClient from "../mqtt";
 
 const LOCK_NAME = "onsite-cart-update";
@@ -52,7 +52,7 @@ export class CartManager {
 
   private async makeRequest<T>(
     input: string | URL,
-    init?: RequestInit
+    init?: RequestInit,
   ): Promise<FallibleRequest<T>> {
     const perform = async () => {
       console.debug("Making request", input);
@@ -94,7 +94,7 @@ export class CartManager {
 
   public async refreshCart() {
     const data = await this.makeRequest<CartResponse>(
-      this.urls.onsite_admin_cart
+      this.urls.onsite_admin_cart,
     );
 
     if (!data.success) {
@@ -126,11 +126,11 @@ export class CartManager {
   public async applyCashPayment(
     reference: string,
     total: string,
-    tendered: string
+    tendered: string,
   ): Promise<FallibleRequest<void>> {
     let url = new URL(
       this.urls.complete_cash_transaction,
-      window.location.href
+      window.location.href,
     );
     url.searchParams.set("reference", reference);
     url.searchParams.set("total", total);
@@ -144,7 +144,7 @@ export class CartManager {
   }
 
   public async enableCardPayment(
-    fallback: boolean
+    fallback: boolean,
   ): Promise<FallibleRequest<void>> {
     let url = new URL(this.urls.enable_payment, window.location.href);
     if (fallback) url.searchParams.set("fallback", "true");
@@ -156,7 +156,7 @@ export class CartManager {
     ids: number[],
     clearCart: boolean = true,
     mqttPrint: boolean = false,
-    beforeClearingCart?: () => void
+    beforeClearingCart?: () => void,
   ): Promise<FallibleRequest<BadgePrintResponse>> {
     const assignData = await this.makeRequest(this.urls.assign_badge_number, {
       method: "POST",
@@ -165,7 +165,7 @@ export class CartManager {
           return {
             id,
           };
-        })
+        }),
       ),
     });
 
@@ -185,7 +185,7 @@ export class CartManager {
         JSON.stringify({
           action: "print",
           url,
-        })
+        }),
       );
 
       if (clearCart) {
@@ -209,7 +209,7 @@ export class CartManager {
   public urlForBadge(id: number): string {
     let url = new URL(
       this.urls.registration_badge_change,
-      window.location.href
+      window.location.href,
     );
     url.pathname = url.pathname.replace("0", id.toString());
     return url.toString();
@@ -222,7 +222,7 @@ export class CartManager {
   }
 
   public async createAndApplyDiscount(
-    amount: string
+    amount: string,
   ): Promise<FallibleRequest<void>> {
     const formData = new FormData();
     formData.set("amount", amount);
@@ -240,7 +240,7 @@ export class CartManager {
 
     let url = new URL(this.urls.onsite_print_receipts, window.location.href);
     this.cartEntries()?.result?.forEach((badge) =>
-      url.searchParams.append("reference", badge.reference)
+      url.searchParams.append("reference", badge.reference),
     );
 
     return await this.makeRequest(url);
@@ -253,7 +253,7 @@ export class CartManager {
 
     let url = new URL(
       this.urls.onsite_admin_transfer_cart,
-      window.location.href
+      window.location.href,
     );
     url.searchParams.append("terminal_id", terminal_id.toString());
     this.cartEntries()?.result?.forEach((badge) => {
