+++ registration/frontend/src/admin/cart/index.ts
@@ -1,4 +1,4 @@
-import { CartManager, CartResponse } from "./cart-manager";
+import { CartManager, type CartResponse } from "./cart-manager";
 import { Cart } from "./components/Cart";
 
-export { Cart, CartManager, CartResponse };
+export { Cart, CartManager, type CartResponse };
