+++ registration/frontend/src/admin/scan/components/UrlEntry.tsx
@@ -1,4 +1,4 @@
-import { Component } from "solid-js";
+import { type Component } from "solid-js";
 
 import { CloseButton } from "./CloseButton";
